Bug fixes in resmgr & fix role for admin user
Resource manager fixes:
1. There is a scenario where the backbone master could report a "new" host as
missing. The poll logic in resource manager did not handle this case well. With
this change, if we get a new host reporting its status as missing, there is
nothing to process - so just ignore it.

2. When we upgrade a host role to a new version, resource manager was not
cleaning up the previous host:role version association in the database. This led
resource manager to report the role multiple times. The UI would show this role
multiple times too. Now, when a host role association is to be stored in the
database, resource manager will remove previous existing roles with the same
name and add the new role version association to it.

Keystone fixes:
3. When we customize a DU. We create an admin user and add it to the "service"
tenant with an admin role. However, keystone adds the _member_ role by default
to the user. In Alpha3, we want to restrict one role per user per tenant. In
this change, we will remove the _member_ role for the user after the addition of
the admin role.

Testing done:
For (1), directly hit the bbmaster PUT API and created a missing host case. Then
ensured that the poller thread in resource manager ignored this host.
For (2), performed a complete upgrade workflow in the DU and the host
components. Noticed that after upgrading ostackhost and imagelibrary on the
host, there was only one host to role association reported in the DB and the UI.
For (3), customized a new DU and verified that the user that was created had
only the admin role assigned to it.

Reviewed by: leb, rdeuel (https://rbcommons.com/s/platform9/r/868/)

Issues addressed: IAAS-778, IAAS-756, IAAS-838