Reload yum cache if needed when querying hostagent
When the hostagent loads, it queries the installed hostagent version. During
upgrade, there is a period of time when both the versions of the hostagent will
coexist leading to an assert in the old host agent code.

Changed the code to now close the rpmdb handle and reopen it when such a
scenario occurs. This should eventually lead to only the new version of
hostagent getting reported as part of the query.

Testing done: Updated the hostagent and verified that the agent initially
reported multiple version of the agent but eventually reported only the upgraded
version.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/904/)

Issues addressed: IAAS-758