Update only the selected role to active version during role update in resmgr
There was a bug in resmgr that would upgrade all the assigned roles to a host,
even if only one role was being upgraded. The issue was that an active role
config cache was maintained in memory. Each time a role put request was sent for
a host, resmgr was looking at the roles assigned to the host and pushing the
active configuration for all the roles. This led to things like imagelibrary
getting updated when only ostackhost update was being attempted.

The cache by itself was not very helpful. It is not the case that host role
assignments are too frequent to we need a cache to optimize it. In this change,
the cache is removed altogether. Instead, resmgr will rely on querying the DB to
get the necessary information. Also, the configuration sent to backbone is
queried from the DB so that it is relies on the real source of truth (i.e. the
DB). This ensures that when only one role is updated, the other one remains
unchanged in the request sent to backbone.

Testing done: Performed role authorization from the DU. Then ran multiple
iterations of updating the 2 roles on the host with a time lag between the 2
role updates. Noticed that only the updated role was changed on the host always.

Reviewed by: leb, osoriano (https://rbcommons.com/s/platform9/r/931/)

Issues addressed: IAAS-816