Fix hostagent python dependency (IAAS-905) and refactor makefiles
Apply the 'no-global-site-packages.txt' trick from ostackhost in
order to create an isolated virtualenv that can still consume
global packages ('yum' on redhat, and perhaps 'apt' on ubuntu)

There was a lot of duplication between the debian and redhat
makefiles, so I refactored the common portion.

Testing: built redhat tarball RPM, and verified contents of both
the RPM and the embedded tarball. In particular, made sure it
contains the 'requests' python module. Will try to test the
Ubuntu version.