Add glance cleanup to janitor service
With this change, the janitor checks for removal of imagelibrary
hosts, and removes the corresponting imagelibrary entries as well
as the images themselves. Also refactored a bit so I could use some
if the code from the nova cleanup.

testing: added an imagelib host, added images (both snap and discovered),
removed the host using the resmgr rest api and verified that images
are deleted.