Fix versioning for the hostagent deb package.
The discover_vms test checks the version of the hostagent package.
For the deb package, the version also should include the release number
and the githash. The deb packages for ostackhost and imagelibrary should
be fine since they already have the version and release included in their
deb versions.

Also, the deb package customization script is now idempotent, since it
does not throw an error if the package already exists.

Testing done:
    * Upgraded an Ubuntu host using the pf9-du-tools, and ensured
      there were no errors.
    * Verified the deb package versions before and after the upgrade.
    * Ran 46-customize-hostagent-deb and ensured no errors when it
      overwrote the existing deb package