Add host roles removal to du-tools
It is helpful to have a way to reset the host roles given that we cannot remove
imagelibrary role from the UI. Adding this the du_ctl script to allow this
functionality during customer support requests.

Exposing 2 commands new commands now.
1. ./du_ctl host list: Will list the hosts and the roles from resource manager.
This can be used to figure out which host's roles need to be reset
2. /du_ctl host role-reset: Takes the host_id that needs to be reset.

The username, password to log in can be provided at the command line for each
command or it can be exported as the environment variable OS_USERNAME and
OS_PASSWORD.

Testing done: Performed host list and host role-reset commands on the DU.

Reviewed by: madhura (https://rbcommons.com/s/platform9/r/1151/)