Fix hostagent init script for Debian-based distros
The previous implementation failed to report the correct status
when hostagent crashed. Also, it threw an warning if the
/var/lock/subsys directory does not exist. This update addresses
those issues.

The start-stop-daemon command will now be used to launch hostagent.
Lock files will not be created anymore. IAAS-1150 will be kept to
track whether or not lock files are needed.

Issues adressed: IAAS-936

Testing done:
* Tried many start/stop/restart operations.
* Crashed hostagent manually when it was up. Ensured the status
  call behaved properly before and after this crash.

Reviewed by: Bich Le (rbcommons.com/s/platform9/r/1159/)