Use getfqdn to get the hostname in hostagent
Use socket.getfqdn to get the hostname in hostagent. The current platform.node
call may or may not return the FQDN of the host. But in nova, libvirt tries to
figure the FQDN of the host. Let's make both of these similar in implementation.

Testing done: Ensured FQDN is reported by hostagent to the bbmaster.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/1215/)

Issues addressed: IAAS-1216