Fix memory leak in hostagent for Ubuntu
There seems to be unusually huge memory usage over a period of time when
apt.cache.update is invoked. Considering that hostagent queries the apt DB
frequently, this is leading to a substantial growth in memory usage over time.
This change avoids the update call during the query operations.  There is
nothing in the documentation that seems to indicate our usage of the update call
is incorrect and the inferences here are just from trial and error. It looks
like this cache update call is required only during install and remove
operations. During query, only a cache open is sufficient.

Caveats:
1. There is still a memory leak (~2-4M) but it is now limited to removal of an
ostackhost or imagelibrary deb. I presume this is not a very common operation at
the customer site.
2. There is a fd leak when we install ostackhost or imagelibrary. Again,
presuming this is not a common operation.

Testing done: Tested authorizing/deuathorizing a host multiple times. Ensured
the memory footprint doesn't grow when the agent is doing only query operations.

Reviewed by: leb, osoriano (https://rbcommons.com/s/platform9/r/1219/)