Do not overwrite values in resmgr add_role API
The body of the PUT resmgr/v1/hosts/<host_id>/roles/<role_name> endpoint
must now be either empty or a Json dictionary. A request with an
empty body is equivalent to a request with an empty Json dictionary. The dictionary specified
by the body must provide a subset of the custom settings.
Here is how the specified Json dictionary is processed.

1. The host will use the custom values specified by the dictionary.
   This will overwrite an existing custom value if it is specified in the body.
2. If the host is being upgraded and if a custom setting was not specified by the body,
   the host will keep the value it had for that custom setting before the call.
3. If this is the first time applying the role to the host and if a custom setting was not
   specified by the body, then the host will get the default value for that custom setting.

Previously, if a host was already at the latest version of a role, another PUT request would
not modify the custom settings for that host (except in the case of the pf9-ostackhost-vmw role).
Now the PUT request will update the custom settings using the steps described above, regardless of
the role being PUT.

Testing done:
   Added/Upgraded a role manually
   Tested changing the custom values of an existing role without deauthorizing

TODOs: Upgrade test changes will be sent out in another review once the test stabilizes

Issues: IAAS-2259

Reviewers: leb, pshanbhag (https://rbcommons.com/s/platform9/r/1294/)