Add token decoder middleware for agent download paste pipeline
The middleware takes in a base64 encoded compressed auth token. It decodes it,
decompresses the token, attaches it to the request header and passes it down the
paste pipeline.

This is implemented only for the static file pipeline. Since only the hostagent
are downloaded from this endpoint, there is an assumption that all requests will
send a encoded compressed token in the request.

Testing done: Tried it with the new UI. Used CLI to download packages. Ensured
negative cases return Unauthorized request.

Reviewed by: rdeuel (https://rbcommons.com/s/platform9/r/1399/)

Issues addressed: IAAS-1629