Fix security bug in bbmaster (IAAS-1665)
The code was using assert with the assumption that python is invoked
with -O option. I was able to crash bbmaster's I/O thread by sending
a message with a bogus opcode.

The fix involves two lines of defense: catch unhandled exceptions
from the top of the IO thread, and check the opcode.