Check file size of downloaded app file
While downloading a pf9app file from the DU, retrieve the content-length from
the response. This should be the size of the file to be downloaded. Proceed to
download the file and eventually check to see if the actual file size matches
the content length. If they don't match, mark as download failure and retry the
convergence logic.

Testing done: Tested authorizing a host - succeeded with the file size check.
Created an error case programmatically in the changed code and verified
convergence eventually went to the failed state.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/1425/)

Issues addressed: IAAS-1641