Substitute rabbit credentials based on app name instead of role name
The recent RabbitMQ changes assumed the app_name matches the role_name.
This is not the case, and it is possible for one role to supply many apps.

Therefore, the rabbit credentials for a specific app configuration will be
substituted by looking up the credentials for the role that provides the
app.

In the future, we might want rabbit credentials for every app in a role.

Review: https://rbcommons.com/s/platform9/r/1501/

Issues: IAAS-1829