Make resmgr wait for local mysqld if it is present
Make resmgr wait for local mysqld if it is present by having the init.d script
specify an optional start requirement. This change uses the Should-Start clause
which will try to run after mysqld if it is present on the machine.

This differs from the Required-Start clause because it will still start
pf9-resmgr even if the mysqld service is not provided by any local service.

This is needed now since we are deploying production deployments with local
MySQL servers as well.

Issues addressed: IAAS-436. This is not a complete fix for this issue but it
does ease up the failure conditions we are hitting.

Testing done: Multiple reboots of EC2 DU and verified that resmgr starts cleanly
after local mysqld. Stopped mysqld service and verified that the resmgr service
tries to come up.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/1519/)