Close connection when io_loop exits
When Pika encounters an error while in bbmaster's io_loop, the
connection and channel may not close properly.  Hence, when
retrying the io_loop, access to the exclusive queue (declared in
the previous io_loop) will be refused.

The connection will now be closed explicitly when the io_loop exits,
and, in the case of an IndexError, a message will be posted to
Slack (5 messages max). Only dev DUs will post to Slack.

Testing done:
  Manually threw an error in send_pending_msgs.
  Ensured the behavior matched the description of IAAS-1909.
  Ensured that, after applying the changes, bbmaster is able to recover from such errors.
  Ensured Slack behavior was correct

Issues: IAAS-1909

Review: https://rbcommons.com/s/platform9/r/1603/