IAAS-1963 - VM's not discovered after deauthorization & reauthorization
Reverted code to remove instances when a hypervisor is deauthorized in
the janitor. (Code added as part of enhancement to janitor IAAS-1917)

Issue: When the hypervisor is deauthorized, the compute node is marked as
deleted in the compute_nodes table. Code was added for janitor to clean
the VM's present in the database when the node was removed. But it also
seems to send out a message to nova-compute to delete the VM. This
message is placed in rabbit's unacknowledged message queue for that compute
node. When the node is re-authorized, ostackhost comes up and consumes this
message and deletes the VM. Thus there are no VM's when the discovery module
starts.
	Going ahead, we will need to change the code to remove the
entires from the database alone and not initiate a delete on the compute
node.