Increase pika timeout for hostagent and rabbit broker connection
For hosts that are geographically further from its DU, we are having issues with
hostagent not being able to establish a SSL connection to the rabbit broker.
This is because pika has by default a 0.25 second timeout on its socket
connection. This is not sufficient for these hosts.

In this change, the code now explicitly set the rabbitmq connection timeout. It
can be provided in the hostagent configuration file otherwise it will be set to
2.5 seconds.

Testing done: Manually installed the new hostagent on a host in Singapore and
Ireland. It was able to connect to its DU in Norther California.

Issues addressed: IAAS-2205

Reviewed by: leb (https://rbcommons.com/s/platform9/r/1764/)