Add openport.py to hostagent
As a short term solution to our port opening needs
and to avoid code duplication, we now place openport.py
to /opt/pf9/hostagent/bin/ and let each Pf9App call this
script in their post-install steps if they want to open ports.

+ Add iptables-persistent as a dependency to debian hostagent.
    We need this to persist iptables rules on Ubuntu machines.
+ Add logic to save iptables rules using iptables-persistent
    in line 40-41 of openport.py

This change is part of IAAS-1318

Testing:
   Built hostagent tarball and verified that openport.py is
   placed in /opt/pf9/hostagent/bin/openport.py

Note:
Not everybody is comfortable with opening customer ports so
we are going to be transparent with which ports we are opening
when a customer/end user authorizes their hosts for the first time.

The idea is to give them a warning/notice in the UI that authorizing
hosts to a particular role will open ports to enable services such
as novncproxy service and imagelibrary.

Review: https://rbcommons.com/s/platform9/r/1798/