Fix paste version used in resmgr
Paste version in pypi was updated yesterday. This is causing resmgr token
authorization code to break. All resmgr calls are leading to bad user token
errors.

This change fixes the paste version to 1.7.5.1 and looks like resmgr works fine
with this version. This should unblock the product while we figure out why the
latest version of Paste is incompatible.

Testing done: Fixed the virtual env in the DU with older paste version and
verified resmgr is able to service requests successfully. Resmgr rpm build was
successful.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/1835/)

Issues addressed: IAAS-2407