Always declare queue when sending to slaves
Currently, we only declare the bbslave queue the first time we send a
message to a particular slave. This causes problems if the queue
is deleted or the broker is restared.

This change will declare the queue every time a message is sent
to a slave.

Testing done:
    Created a DU with the changes.
    Ensured that after a broker restart, the slave is still able to receive messages
    Ensured that after the queue is deleted, the slave is still able to receive messages

Reviewers: leb (https://rbcommons.com/s/platform9/r/1836/)

Issues: IAAS-2363