Avoid using systemctl in hostagent RHEL init script
For CentOS, we use the init functions in /etc/init.d/functions
for start/stop/status operations. On CentOS 7, if the
SYSTEMCTL_SKIP_REDIRECT environment variable is empty, then
the start/stop/status operations calls are forwarded to systemctl.

Since our init scripts are not compatible with systemctl, we need
this environment variable for CentOS 7.

Issues: IAAS-2382

Reviewers: pshanbhag (https://rbcommons.com/s/platform9/r/1904/)