Fix single role removal from host in resource manager
The existing code was doing most of the things correctly, except the case where
the role being deleted was the last role in the host. With this change, check if
the role is the last role and in such cases perform a host delete operation.
Also, introduced a role delete lock to prevent race conditions when we determine
if this is the last role or not.

Testing done: In a DU, performed the resmgr delete role API calls from the CLI
and ensured role removal happens as expected. Verified with contents in DB.

Issues addressed: IAAS-2396, IAAS-649

Reviewed by: osoriano (https://rbcommons.com/s/platform9/r/1941/)