Use only host assigned roles for convergence check
When checking if the host configuration is in convergence, resmgr was looking at
the all the roles that are active. This is not valid always, especially when the
host roles are not yet updated to the latest versions. In such cases, resmgr
would end up trying to push new configuration to the host, using the latest host
role data. This would lead to comms being pushed down as well by bbmaster\!
During that process, hostagent connectivity is briefly lost. This could also
lead to other upgrade messages (like hostagent upgrade) getting lost, breaking
the upgrades.

In this change, we shall use only the roles assigned to the host to build the
expected configuration when determining host convergence. This should eliminate
the case where we push new configuration when newer role versions are available.

There are more optimizations one can do here, esp. simplify
_update_custom_role_settings method. Will be done as of another change.

Issues addressed: IAAS-2609

Testing done: Set up a DU with these changes. Ensured resmgr authorize,
deauthorize works fine. Also, no issues with the running of the backbone poller
loop. Ran upgrade test.

Reviewed by: akapoor, osoriano (https://rbcommons.com/s/platform9/r/1959/)