bbmaster changes for cluster setup
When rabbitmq is setup in a cluster, the bbmaster on all the nodes should get a
copy of the messages published by the hosts. In order to achieve this, we will
continue to use a direct exchange. We create a separate bbmaster queue for each
node, but all of the queues will bind to the same existing routing key. This
makes it behave like a fanout exchange. Hostagent publishes messages with this
same routing key and it should get delivered to all the queues.

This change is one of many in achieving DU clustering capability.

Related review: https://rbcommons.com/s/platform9/r/2015/
Issues addressed: IAAS-2719
Reviewed by: osoriano