Add a role-specific config mechanism to the resource manager
This change gives resource manager the ability to read configuration
ini files from a directory on the DU and make the configuration
available for rendering role json strings.

For some of the newer roles that have been added, we've hardcoded a
lot of configuration values in the resmgr's dbutils module, and
added them to the global.conf file. To make things more modular,
both in the resmgr as well as in the ansible scripts that setup the
initial role configs, I've added a configuration directory (conf.d)
to the resmgr_roles path, into which ansible or other role setup code
can deposit ini files. Here's an example of a role config file:

[pf9-glance]
sqlpasswd = 7avQGT5PAVZb4b7x
auth_user = glance
auth_pass = I4UVC2e1Uk1lhl6W
auth_tenant_name = services

This file is created when we customize the DU with ansible. Resource
manager reads this (and other files in conf.d) and adds the values to
its ConfigParser object. When a role is configured, the entire config
is made available as a flattened dictionary with names representing
the section.value names from the config. For example, sqlpasswd above
can be rendered into the glance role json with %(pf9-glance.sqlpasswd)s.

This removes this config from global.conf, and allows us to move
the ansible configuraton for each role into its own module. Also, it
removes the need for dbutils to be aware of specific config values.

testing: deployed new glance roles and verified correct configuration,
both in the roles database table as well as in the host's glance
config.