Add DU-side role configuration mechanism to resmgr
For the new on-host glance-api role, we need the ability to add
a glance endpoint to keystone when the role is added, and remove
it when the role is removed. This change adds a mechanism to the
resource manager to run role-specific code at role authorization
and deauthorization time in the DU. The implementation for the
event methods (on_auth and on_deauth) are provided by the role
wrapper along with the role configuration json document. To activate
the feature, the role wrapper must provide markup like the following
example in the json config:

...
"customizable_settings": {
    "glance_address": {
        "path": "du_config/auth_events/params",
        "default": ""
    }
},
"config": {
    "pf9-glance-role": {
        "version": "__VERSION__",
        "service_states": {...},
        "url": "...",
        "du_config": {
            "auth_events": {
                "type": "python",
                "module_path": "/etc/pf9/resmgr_roles/pf9-glance-role/__VERSION__/auth_events.py",
                "params": {
                    "admin_user": "%(pf9-glance.auth_user)s",
                    "admin_pass": "%(pf9-glance.auth_pass)s",
                    "admin_url": "https://localhost:35357/v2.0"
                }
            }
        },
        "config": {
            "glance_api_conf": {...}
...

In the above, the 'du_config' dictionary tells the resource manager
where to find the event implementations and how to run them:
+ auth_events.py should contain on_auth and on_deauth functions
+ parameters are provided as **kwargs to the methods. The above
  du_config spec will cause resmgr to provide 4 keyword parameters
  when calling the event methods:
  + admin_user - rendered into the json config from the resmgr's
                 pf9-glance.auth_user configuration param
  + admin_pass - like admin_user
  + admin_url -  hardcoded in the role.json
  + glance_address - a 'customizable setting' passed into resmgr at
                 authorization-time.

The event methods themselves therefore look like:
def on_auth(glance_address, admin_user, admin_pass, admin_url):
   ...
def on_deauth(glance_address, admin_user, admin_pass, admin_url):
   ...

testing:
Ran this with the new glance role (stubbed event methods), simulating
both success and failure conditions. Also ran as a private build against
atherton to verify this functionality is a no-op on existing roles
that don't contain the du_config keys.