Make firmware apps use the new switcher/comms rearchitecture URL
Make firmware apps use the new switcher/comms rearchitecture URL. We need this
to have the capability to switch the hostagent downloads from https to http and
also use the new ports.

Testing done:
Made changes to bbmaster in catfood and ensured pf9-apps are getting installed.
Prior to this, the hosts were unable to download the firmware apps.

Reviewed by: pacharya, leb (https://rbcommons.com/s/platform9/r/2246/)