Get list of volumes present on host. - IAAS-2137
When enabling LVM driver, the driver requires "cinder-volumes" volume
group to be present on the host. This change will add the list of
volumes present on the host. During authorization, the UI needs to
check if "cinder-volumes" is part of the list of volumes returned.

UI bug for tracking this: IAAS-3308