Fix PYTHONPATH in hostagent when invoking host system modules
With yum and apt, the system python modules should be used and having the
Platform9 PYTHONPATH as part of the environment causes these modules to not load
properly.

The PYTHONPATH should not generally be exposed to the environment running these
package manager commands because of default sudo rules on the host, but that can
be overriden by customers when they customize the sudoers file.

Fix is to drop the custom PYTHONPATH whenever the package manager commands are
invoked. There is an odd behavior with subprocess module here. Once a customized
environment is passed to subprocess module, it seems to stick for subsequent
calls as well if you don't pass in an env setting. The docs don't indicate this
though. I was not successful in root causing what is causing it. To get around
this, I am passing in os.environ as default value to the subprocess calls.

Issues addressed: IAAS-3305

Testing done: Patched a hostagent on Scientific Linux 6 and Ubuntu14 box.
Performed role authorization/deauthorization on each host with PYTHONPATH
exposed to the environment in the sudoers file. Performed the same with the
PYTHONPATH not exposed in sudoers file. Will ensure integration tests pass.

Reviewed by: osoriano (https://rbcommons.com/s/platform9/r/2279/)