pf9-du: changes to support DU-in-containers
The hostagent.conf in the generated host agent RPM now contains
a [du] section containing a flag indicating whether the DU runs
in a container. pf9-comms uses this flag to determine whether
to prefix SNI server names with the FQDN.

The setting of this flag is controlled by the DU_IS_CONTAINER
environment variable that should be passed in during Ansible
configuration phase.

Also delete 05-generate-certs, since it's now done outside of the DU
by code residing in pf9-main.