Add executable mode check for running the extension scripts
Add executable mode check for running the extension scripts. This will skip over
scripts that cannot be run.

Also, added an explicit exception handling for the code to run the extension
command so that an appropriate error is returned as part of the extension data.

Testing done: Patched the code in a host where the error was occurring and
ensured that hostagent is able to deal with them as expected

Reviewed by: leb, osoriano (https://rbcommons.com/s/platform9/r/2389)

Issues addressed: IAAS-3468