IAAS-3374 Fetch free and used volume data
Along with volumes present on the device, now the extension will also
gather and send the "Size of volume" and "Free space" of the volume.
Output:
{u'volumes_present': {u'status': u'ok', u'data':
[{u'name': u'cinder-volumes', u'free': u'12.00g', u'size': u'12.00g'},
 {u'name': u'vg_nestedctos1', u'free': u'0', u'size': u'7.51g'}]}

Will return empty list if no volumes is present.
Eg:
{u'volumes_present': {u'status': u'ok', u'data': []}

Testing done:
Replaced script in a hypervisor and checked the heartbeat logs on
the DU to verify result. Tried it with hypervisor with no cinder-volumes.