Use associated role object when removing roles from host in DB
Use associated role object when removing roles from host in DB. Previously, the
code would look for the active role in the DB and try to remove it. While this
works for most of the usual use cases, it will not if used during an upgrade
like when the DU is upgraded but host is not yet upgraded. This is because there is
a newer role version active and the host is having the older role version.

Now the code will match the name against the role objects associated with the
host and remove the appropriate role.

Testing done: Using curl, invoked role delete command. Verified that the role
deletion happens fine. Will test the upgrade path as part of the upgrade test
fixes.

Reviewed by: osoriano (https://rbcommons.com/s/platform9/r/2457/)