Move auth_events processing to per app instead of per role
The auth events in the roles json are a property of each app. Instead of looking
for this property based on role name, we now will look for it for each app
specified in the role. This will also allow different roles which reuse the same
app to also handle these auth events.

Testing done:
Patched a DU with the changes and authorized the glance role. Ensured the glance
role endpoint gets registered. Will try to test for glance-vmw scenario too.

Reviewed by: rdeuel, susmitha(https://rbcommons.com/s/platform9/r/2549/)