Package latest certs into hostagent
With the recent changes to switcher and comms, there may be multiple
sets of certs on the DU. The convention is that, inside the
/etc/pf9/certs directory, there will be directories such as v1.2.0,
v1.3.0, etc. This change ensures the latest certs will be packaged
into hostagent when building the rpm on the DU.

Note that we must store the cert_version in hostagent.conf since
comms needs to know which server_name to request.

Design: https://platform9.atlassian.net/wiki/display/eng/Certificate+Strategy+in+1.4

Testing done: Passed the build

Issues: IAAS-2723 IAAS-3880