Only run events for apps being added or removed
Previously, events were being called for all apps which leads to errors
if the app version is not present on the DU. Instead, events
will only be run for the apps being added or removed.

This ensures that, when a DU is being upgraded, the on_auth call will
have access to the modules it needs (since it will be provided by new app version).

Testing done:
    Upgrade ostackhost on an imagelibrary. Ensured no imagelibrary events were triggered.

Issues: IAAS-3653