Add app ordering for setting up pf9apps
Introducing ranks for apps to determine the order in which they should be set
up. There is a case where a role can have multiple apps and there could be a
dependency between those apps - needing for an ordering on how they are
installed. The rank of an app is a float (x.y). The reasoning is that all apps
in the same role will share a unique x value but will have different y values.
This should prevent collisions of ranks between different roles. For apps which
don't specify a rank, a default value of 0.0 is assumed both in bbmaster level
as well as the hostagent level.

Testing done: Unit tests pass. Ensured no regressions through private build:
http://teamcity.platform9.sys:8111/viewLog.html?buildId=168979&buildTypeId=Pf9project_Integration_DuAmiIncr&tab=buildResultsDiv

Reviewed by: leb (https://rbcommons.com/s/platform9/r/2769/)