Copy over merge and extract utility methods for ini configs
These methods can be generic and reused by different pf9apps when they set and
get config details.

This is a copy of the utility methods that is present in glance repo's pf9app
config script. There was one extra case that was added in here - the if
condition around line 231

See https://platform9.atlassian.net/browse/IAAS-3810 for adding this into the
unit tests for configutils.

Testing done: Ran DHCP agent pf9app setup with this and it verified get and set
configs work as expected.

Reviewed by: rdeuel (https://rbcommons.com/s/platform9/r/2779/)