Add support for new appliance for the neutron network node
Modified the 47-customize-hostagent-ova script to accept a new arg
neutron_vmw, which if set to true will create a network node appliance.

Verified that if this arg is not supplied or supplied with a different
value, the script does not end in error and the network node appliance
is not created.