IAAS-4081: Neutron: Set vmware hypervisor type in parent host_info object using the value from hypervisor_details file
Currently only existence of hypervisor_details file is used to indicate if hypervisor type is VMware.
But with Neutron, the hypervisor type in hypervisor_info needs to read the value from the file (VMwareCluster or VMwareNeutron).

Also fix bbmaster/pf9_vmw_mgmt to treat hypervisor with type VMWareNeutron as a VMware appliance.