hostagent: enable remote ssh support command that includes public key
This is a stripped-down version of Omar's change at
https://rbcommons.com/s/platform9/r/2827

It only includes the hostagent changes needed to support a new variant
of the remote ssh command. This uses the key pair authentication
mechanism, eliminating the need for the customer to do set up
anything for us.

The integration-test Make target passes.
I will update this review with additional tests.