IAAS-3166: pf9-hostagent ensures pf9-comms always in running state
If pf9-comms is installed but not running, set it to running
state to ensure hostagent can communicate with the DU. This situation can
happen if pf9-hostagent and pf9-comms packages are installed manually
outside of the installer, and pf9-hostagent service is started first.

In that situation, if pf9-hostagent runs for the very first time,
desired_apps.json does not yet exist, and is created based on the
current state of the system. Since pf9-comms is a pf9 app, if it is
not running at that time, its desired state is captured as 'not
running'. Which is bad, since pf9-hostagent will shut it down if it
is ever started.

Tested these 2 scenarios:
- The initial RPM installation scenario (not using installer)
  described above.
- In steady state, manually edited desired_apps.json to set
  pf9-comms' desired run state to False.
In both cases, verified that hostagent logs an entry indicating
IAAS-3166 handling, starts pf9-comms, and updates desired_apps.json
correctly.