Janitor should deal with case where nova requests may return non OK codes
When the nova request fails, the method was returning None which the call site
could not figure out how to unpack. As part of the change, the method will raise
exception in such a case which will be handled at the call site and handled
gracefully.

Issues addressed: IAAS-4121

Testing done: Patched a DU and ensured janitor is OK when nova-api is down.
Restarted nova-api and ensured janitor does its thing

Reviewed by: susmitha (https://rbcommons.com/s/platform9/r/2988/)