support enhancements to hostagent: support command regex + more bundle dirs
To alleviate some of the recent and future upgrade issues,
we are exploring deploying another control agent alongside HA.
This other agent will likely expose an API allowing remote invocation
of the same allowed support commands specified in
/etc/pf9/allowed_commands. This change moves the remaining
commands still hard-coded in HA into the directory.

As part of this, also found and fixed a bug in the way support
commands specified in files were read.

Testing: still testing, and also completed a du-ami-incr build with
this change to ensure it didn't introduce regressions:
http://teamcity.platform9.sys:8111/viewLog.html?buildId=190701&tab=buildResultsDiv&buildTypeId=Pf9project_Integration_DuAmiIncr