resmgr tokendecoder should not fail if the token isn't compressed
Added error handling to the tokendecoder to catch decompress errors
and allow the raw cookie to proceed through to the auth_token
middleware.

In the new multi-DU setup, the UI won't be generating the
cookiefied token anymore. We'll expose an endpoint in each region's
DU that accepts an X-Auth-Token header and generates a cookie
for that host's domain. See
https://platform9.atlassian.net/wiki/display/~rdeuel/2015/12/29/Multi-Region+Endpoints

Also note that with keystone v3, it won't be necessary to compress
tokens. When fetched using /keystone/v3/auth/tokens?nocatalog, the
scoped token comes back without the service catalog, and is a reasonable
size, about 1300 characters. The cookie set from the /token2cookie endpoint
will not be compressed.

testing: tested with both compressed and uncompressed cookies.