Remove comms config chown when upgrading hostagent
Hostagent was recursively chowning files in /etc/pf9 on upgrades.

This results in an issue for cert upgrades since comms is watching
directory /etc/pf9/comms/sni_maps. Comms will think the map has been
modified, so the new certs are used, but comms is not aware that the
cert version has changed.

Testing done:
    Passed the upgrade_certs test and a subset of the integration tests.